==> 5.Genotype_script.R <==
library(data.table)

genotypingSum <- fread("path/RAD-seq/NGS_Data_Analysis_folder/parent/mappinggenotypelist.csv", header=TRUE)
genotype<- c("tags","freq")
dataP1csv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/parent/P1unique_tag1over.csv", header=F, col.names=genotype)
system.time(write.csv(dataP1csv[match(genotypingSum$tags, dataP1csv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/parent/P1.tagcount1overmatchindex.csv", row.names = F))
genotypingP1 <- fread("path/RAD-seq/NGS_Data_Analysis_folder/parent/P1.tagcount1overmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingP1)
dataP2csv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/parent/P2unique_tag1over.csv", header=F, col.names=genotype)
system.time(write.csv(dataP2csv[match(genotypingSum$tags, dataP2csv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/parent/P2.tagcount1overmatchindex.csv", row.names = F))
genotypingP2 <- fread("path/RAD-seq/NGS_Data_Analysis_folder/parent/P2.tagcount1overmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingP2)
write.csv((genotypingSum), file="path/RAD-seq/NGS_Data_Analysis_folder/parent/genotypingSum_P1P2.csv", row.names = F)

datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_001.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_001.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_001.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_002.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_002.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_002.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_003.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_003.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_003.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_004.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_004.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_004.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_005.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_005.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_005.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_006.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_006.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_006.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_007.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_007.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_007.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_008.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_008.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_008.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_009.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_009.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_009.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_010.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_010.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_010.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_011.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_011.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_011.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_012.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_012.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_012.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_013.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_013.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_013.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_014.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_014.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_014.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_015.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_015.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_015.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_016.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_016.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_016.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_017.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_017.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_017.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_018.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_018.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_018.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_019.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_019.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_019.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_020.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_020.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_020.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_021.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_021.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_021.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_022.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_022.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_022.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_023.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_023.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_023.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_024.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_024.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_024.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_025.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_025.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_025.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_026.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_026.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_026.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_027.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_027.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_027.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_028.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_028.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_028.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_029.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_029.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_029.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_030.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_030.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_030.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_031.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_031.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_031.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_032.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_032.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_032.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_033.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_033.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_033.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_034.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_034.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_034.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_035.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_035.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_035.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_036.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_036.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_036.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_037.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_037.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_037.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_038.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_038.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_038.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_039.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_039.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_039.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_040.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_040.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_040.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_041.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_041.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_041.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_042.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_042.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_042.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_043.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_043.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_043.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_044.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_044.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_044.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_045.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_045.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_045.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_046.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_046.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_046.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_047.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_047.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_047.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_048.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_048.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_048.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_049.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_049.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_049.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_050.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_050.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_050.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_051.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_051.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_051.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_052.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_052.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_052.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_053.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_053.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_053.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_054.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_054.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_054.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_055.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_055.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_055.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_056.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_056.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_056.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_057.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_057.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_057.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_058.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_058.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_058.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_059.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_059.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_059.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_060.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_060.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_060.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_061.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_061.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_061.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_062.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_062.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_062.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_063.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_063.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_063.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_064.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_064.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_064.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_065.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_065.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_065.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_066.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_066.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_066.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_067.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_067.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_067.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_068.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_068.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_068.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_069.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_069.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_069.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_070.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_070.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_070.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_071.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_071.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_071.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_072.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_072.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_072.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_073.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_073.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_073.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_074.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_074.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_074.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_075.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_075.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_075.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_076.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_076.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_076.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_077.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_077.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_077.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_078.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_078.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_078.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_079.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_079.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_079.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_080.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_080.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_080.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_081.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_081.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_081.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_082.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_082.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_082.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_083.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_083.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_083.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_084.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_084.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_084.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_085.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_085.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_085.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_086.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_086.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_086.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_087.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_087.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_087.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_088.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_088.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_088.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_089.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_089.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_089.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_090.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_090.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_090.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_091.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_091.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_091.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_092.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_092.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_092.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_093.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_093.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_093.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_094.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_094.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_094.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_095.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_095.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_095.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_096.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_096.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_096.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_097.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_097.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_097.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_098.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_098.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_098.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_099.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_099.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_099.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_100.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_100.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_100.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_101.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_101.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_101.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_102.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_102.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_102.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_103.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_103.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_103.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_104.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_104.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_104.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_105.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_105.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_105.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_106.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_106.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_106.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_107.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_107.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_107.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_108.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_108.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_108.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_109.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_109.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_109.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_110.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_110.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_110.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_111.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_111.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_111.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_112.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_112.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_112.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_113.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_113.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_113.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_114.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_114.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_114.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_115.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_115.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_115.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_116.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_116.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_116.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_117.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_117.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_117.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_118.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_118.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_118.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_119.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_119.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_119.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_120.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_120.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_120.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_121.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_121.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_121.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_122.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_122.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_122.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_123.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_123.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_123.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_124.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_124.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_124.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_125.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_125.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_125.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_126.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_126.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_126.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_127.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_127.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_127.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_128.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_128.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_128.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_129.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_129.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_129.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_130.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_130.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_130.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_131.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_131.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_131.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_132.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_132.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_132.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_133.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_133.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_133.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_134.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_134.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_134.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_135.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_135.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_135.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_136.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_136.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_136.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_137.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_137.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_137.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_138.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_138.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_138.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_139.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_139.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_139.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_140.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_140.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_140.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_141.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_141.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_141.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_142.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_142.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_142.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_143.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_143.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_143.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_144.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_144.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_144.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_145.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_145.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_145.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_146.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_146.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_146.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_147.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_147.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_147.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_148.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_148.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_148.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_149.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_149.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_149.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_150.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_150.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_150.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_151.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_151.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_151.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_152.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_152.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_152.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_153.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_153.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_153.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_154.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_154.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_154.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_155.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_155.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_155.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_156.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_156.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_156.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_157.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_157.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_157.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_158.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_158.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_158.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_159.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_159.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_159.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_160.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_160.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_160.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_161.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_161.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_161.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_162.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_162.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_162.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_163.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_163.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_163.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_164.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_164.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_164.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_165.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_165.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_165.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_166.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_166.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_166.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_167.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_167.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_167.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_168.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_168.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_168.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_169.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_169.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_169.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_170.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_170.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_170.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_171.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_171.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_171.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_172.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_172.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_172.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_173.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_173.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_173.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_174.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_174.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_174.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_175.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_175.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_175.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_176.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_176.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_176.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_177.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_177.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_177.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_178.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_178.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_178.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_179.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_179.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_179.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_180.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_180.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_180.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_181.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_181.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_181.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_182.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_182.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_182.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_183.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_183.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_183.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_184.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_184.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_184.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_185.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_185.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_185.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_186.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_186.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_186.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_187.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_187.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_187.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_188.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_188.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_188.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_189.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_189.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_189.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_190.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_190.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_190.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_191.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_191.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_191.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_192.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_192.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_192.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_193.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_193.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_193.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_194.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_194.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_194.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_195.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_195.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_195.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_196.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_196.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_196.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_197.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_197.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_197.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_198.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_198.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_198.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_199.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_199.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_199.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)
datacsv <- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_200.tagcount", header=F, col.names=genotype)
system.time(write.csv(datacsv[match(genotypingSum$tags, datacsv$tags),2], file="path/RAD-seq/NGS_Data_Analysis_folder/F2_200.tagcountmatchindex.csv", row.names = F))
genotypingdata<- fread("path/RAD-seq/NGS_Data_Analysis_folder/F2_200.tagcountmatchindex.csv", header = T)
genotypingSum <- cbind(genotypingSum, genotypingdata)

genotypingSum[is.na(genotypingSum)] <- 0
write.csv((genotypingSum), file="path/RAD-seq/NGS_Data_Analysis_folder/genotypingSum_complete.csv", row.names = F)
Num0 <- sum(genotypingSum[1, 9:ncol(genotypingSum)] == 0)
Num0over <- sum(genotypingSum[1, 9:ncol(genotypingSum)] > 0)
AllNum <- Num0 + Num0over
Num1 <- c(Num0, Num0over, AllNum)
Num0 <- sum(genotypingSum[2, 9:ncol(genotypingSum)] == 0)
Num0over <- sum(genotypingSum[2, 9:ncol(genotypingSum)] > 0)
AllNum <- Num0 + Num0over
Num2 <- c(Num0, Num0over, AllNum)
Nums <- rbind(Num1, Num2)
colnames(Nums) <- c("0", ">0", "All")
for(i in 3:nrow(genotypingSum)){
Num0 <- sum(genotypingSum[i, 9:ncol(genotypingSum)] == 0)
Num0over <- sum(genotypingSum[i, 9:ncol(genotypingSum)] > 0)
AllNum <- Num0 + Num0over
Num3over <- c(Num0, Num0over, AllNum)
Nums <- rbind(Nums, Num3over)
}
Chitest1 <- chisq.test(c(Nums[1,1], Nums[1,2]), p=c(25, 75)/100, correct = FALSE)
Chitest2 <- chisq.test(c(Nums[2,1], Nums[2,2]), p=c(25, 75)/100, correct = FALSE)
Chitests <- rbind(Chitest1$p.value, Chitest1$p.value)
for(o in 3:nrow(Nums)){
Chitest3over <- chisq.test(c(Nums[o,1], Nums[o,2]), p=c(25, 75)/100, correct = FALSE)
Chitest3over
Chitests <- rbind(Chitests, Chitest3over$p.value)
}
colnames(Chitests) <- "Chisquare"
ChiNums <- cbind(Chitests, Nums)
ChiNums <- cbind(ChiNums, genotypingSum)
Genotype <- subset(ChiNums, ChiNums$Chisquare > 0.001)

write.csv(Genotype, "path/RAD-seq/NGS_Data_Analysis_folder/Genotype.csv", row.names = FALSE)
