==> 7.Post_data_correction_script.R <==
library(ABHgenotypeR)
genotype <- readABHgenotypes("path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs.csv", "P1", "P2", readPos=TRUE)

ErrCorr1 <- correctStretches(genotype, maxHapLength = 1)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 1)
SUErrCorr3_1 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_1, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR1SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 2)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 2)
SUErrCorr3_2 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_2, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR2SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 3)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 3)
SUErrCorr3_3 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_3, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR3SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 4)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 4)
SUErrCorr3_4 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_4, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR4SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 5)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 5)
SUErrCorr3_5 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_5, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR5SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 6)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 6)
SUErrCorr3_6 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_6, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR6SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 7)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 7)
SUErrCorr3_7 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_7, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR7SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 8)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 8)
SUErrCorr3_8 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_8, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR8SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 9)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 9)
SUErrCorr3_9 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_9, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR9SU.csv")

ErrCorr1 <- correctStretches(genotype, maxHapLength = 10)
ErrCorr2 <- correctUndercalledHets(ErrCorr1, maxHapLength = 10)
SUErrCorr3_10 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(SUErrCorr3_10, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR10SU.csv")


ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 1)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 1)
USErrCorr3_1 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_1, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR1US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 2)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 2)
USErrCorr3_2 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_2, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR2US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 3)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 3)
USErrCorr3_3 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_3, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR3US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 4)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 4)
USErrCorr3_4 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_4, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR4US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 5)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 5)
USErrCorr3_5 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_5, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR5US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 6)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 6)
USErrCorr3_6 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_6, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR6US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 7)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 7)
USErrCorr3_7 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_7, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR7US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 8)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 8)
USErrCorr3_8 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_8, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR8US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 9)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 9)
USErrCorr3_9 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_9, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR9US.csv")

ErrCorr1 <- correctUndercalledHets(genotype, maxHapLength = 10)
ErrCorr2 <- correctStretches(ErrCorr1, maxHapLength = 10)
USErrCorr3_10 <- imputeByFlanks(ErrCorr2)
writeABHgenotypes(USErrCorr3_10, outfile = "path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR10US.csv")

pdf("path/RAD-seq/NGS_Data_Analysis_folder/Genotype_csvs_ABHgenotypeR1-10SUUS.pdf")
plotGenos(genotype)
plotGenos(SUErrCorr3_1)
plotGenos(SUErrCorr3_2)
plotGenos(SUErrCorr3_3)
plotGenos(SUErrCorr3_4)
plotGenos(SUErrCorr3_5)
plotGenos(SUErrCorr3_6)
plotGenos(SUErrCorr3_7)
plotGenos(SUErrCorr3_8)
plotGenos(SUErrCorr3_9)
plotGenos(SUErrCorr3_10)
plotGenos(USErrCorr3_1)
plotGenos(USErrCorr3_2)
plotGenos(USErrCorr3_3)
plotGenos(USErrCorr3_4)
plotGenos(USErrCorr3_5)
plotGenos(USErrCorr3_6)
plotGenos(USErrCorr3_7)
plotGenos(USErrCorr3_8)
plotGenos(USErrCorr3_9)
plotGenos(USErrCorr3_10)
dev.off()
